CRAYSAT-1038: Remove step to check config import job
This documentation change accompanies an enhancement to the installer
script to automatically wait for the configuration import job to complete.
It is no longer necessary to check the status of the
sat-config-import-x.y.z job because the script does it automatically.

Additionally, the example output from the install script, `install.sh` is
updated, and the step to check the return code is now removed, since
a confirmation message has been added to the install script.