STP-2844: Redo logic for sat-versioning.sh
This changes the logic from sorting the release branches and choosing
the latest to always using whatever branch we are currently on as the
branch of sat-product-stream to use.

Additionally, this change removes sat-version.txt from the repository
once again and adds it to .gitignore, so that stale versions can not
be used. Finally, modify the Jenkinsfile slightly to call the SAT
versioning script, so that setting STREAM_VERSION finds the
sat-version.txt file.

Remove some REPO_PATH variables where they are not needed.

Test Description: After pushing this branch, I verified that the
Jenkins job built sat docs with a version of UNKNOWN. Then, I
temporarily pushed a branch of sat-product-stream with the same name
as this branch, and verified that the Jenkins job built sat docs with
the same version as the branch of sat-product-stream had.