CRAYSAT-1223: Strip duplicate license notice from generated .md
We have been asked to provide the Markdown version of our documentation.
Currently when our `.mds` file is compiled, and `_include` directives
are resolved, we end up with the license text from every included file
being duplicated in the compiled file, which makes the file less
readable.

Strip the duplicated license notice that results from including multiple
files with the `_include` directive that all individually include the
license text. Leave the license text in place for the top-level `.mds`
file, but strip it from all files that are being included.

Test Description:
Ran the following to build the `.md` file with this change in place:

    cd portal/developer-portal
    ../../scripts/compileMDS.py HPE_Cray_EX_System_Admin_Toolkit_Guide.mds

Ran the above again without the change in place and compared the two
results, ensuring the only change was the removal of the duplicated
license text from the compiled `.md` file.