CRAYSAT-1322: Add documentation for `sat bootprep`
This change adds a section to the System Admin Toolkit Guide describing
how to use the `sat bootprep` system. It includes a brief tutorial on
writing bootprep input files, an example bootprep input file,
information on how to generate custom example files, and instructions on
viewing generated documentation and the man page.

Test Description:
Build both HTML and PDF copies of the documentation and inspect for
formatting. Proof read for spelling and grammar.