CRAYSAT-1329: Document creating CFS configuration during fresh install
During a fresh install, the admin will be constructing a CFS
configuration in a JSON file and will not necessarily have created it
yet in CFS. As a result, the automation implemented in `cfs-config-util`
and called by the SAT `install.sh` script will not find any CFS
configurations to modify. However, it will now print out the SAT layer
in JSON format, so the admin can manually add this layer to the CFS
configuration JSON file that is in progress. Document how to handle this
case.

Test Description:
Manually executed any commands that are included as examples here.
Separately tested the changes to `cfs-config-util` that result in output
of the SAT layer as JSON.

(cherry picked from commit 706e7c3298d3ac67a42bc7aeb187a9547e4005c4)