CRAYSAT-1351: Document how to get the correct SAT config layer name
Add a step to save the new name of the NCN personalization CFS
configuration layer for SAT, and use that saved name in the later step
that creates the CFS session(s).

The layer is now named 'sat-x.y.z' where 'x.y.z' is the version of the
SAT product stream instead of just 'sat-ncn'.

Test Description:
This is essentially what I did while installing the SAT 2.2.16 release
on lemondrop on Friday 2/25.