CRAYSAT-1294: Hard-code version of SAT product in docs
When we query the version of the SAT product stream from the
sat-product-stream repository, it is possible for the approved version
of the sat release distribution and the version specified in the sat
docs tar file to get out of sync. For example, sometimes we build a new
version of the sat product stream on the release branch to test some
build tooling changes without the intent of releasing it. In that case,
the SAT version specified in sat-docs will be ahead of the officially
release SAT version.

Fix this problem by just using a hard-coded major and minor version for
SAT in the docs.

Test Description:
Examined the built docs tar file in arti and confirmed that it contained
the correct major and minor version.