CRAYSAT-1372: Add release notes for SAT 2.2
Create release notes for SAT 2.2. The sources of these release notes are
the `CHANGELOG.md` files in the various component repositories
(primarily hpc-sat and hpc-sat-podman), commit logs for the
hpc-sat-product-stream repository, and Jiras in the CRAYSAT project with
SAT 2.2 as the fix version.

Test Description:
Built and inspected HTML and PDF. Checked for broken links and any
formatting issues and typos.