CRAYSAT-1418: Update NCN personalization instructions
Update the instructions for updating CFS configurations that are applied
to management NCNs to use the new `update-mgmt-ncn-cfs-config.sh`
script included in SAT release distribution.

Refine the "Activate" procedure by referring to the updated steps to
apply the CFS configuration to management NCNs. The procedure is the
same whether it is being performed after the install/upgrade of a new
version of SAT or after the activation of a version of SAT.

Test Description:
Build a SAT 2.4.2 release distribution from the corresponding
CRAYSAT-1418 branch of sat-product-stream and installed on baldar.

Followed all three alternatives in the new NCN personalization
procedure and ensured they worked as documented.

Co-authored-by: Benjamin Rohlfing <91636159+Benjamin-Rohlfing-HPE@users.noreply.github.com>
Co-authored-by: Eli Kamin <77810337+ekamin-hpe@users.noreply.github.com>