CRAYSAT-1417: SAT documentation RPM
This commit adds an RPM build of the SAT documentation.

Add a Makefile, RPM spec file, and version scripts  all used for the
building of the docs-sat RPM. The RPM will install all the markdown
files in the docs/ directory of this repository into
/usr/share/doc/sat/, except that the top-level README.md is
now the concatenation of two files, rpm/README.md and docs/README.md
(the latter being the usual README.md), so that some information
specific to the offline documentation can be provided.

There are two version scripts, 'version.sh' and 'commit_count.sh'
that are added. They use git tags to determine the major/minor
version (e.g. 2.4) and the number of commits since the tag defining
that version. These are provided as environment variables in the
Jenkinsfile. The version script is also referenced in the Makefile
and spec file.

This commit adds Jenkinsfile.github for building the RPM in a Jenkins
pipeline. The RPM is published to one of two locations depending on
whether the build is stable. For stable builds, the RPM is published
to: "sat-rpms/hpe/stable/sle-15sp3/docs-sat/X.Y/noarch/" where X and Y
are the major/minor version numbers of SAT (e.g. 2.4). Additionally,
when publishing a stable RPM, a version named
"docs-sat-latest.noarch.rpm" is published alongside the usual
"docs-sat-X.Y-COMMIT_COUNT.noarch.rpm", in the same directory.

For unstable builds (e.g., development branches), they are published
to "sat-rpms/hpe/unstable/sle-15sp3/docs-sat/noarch" (note the
"unstable" directory and lack of major.minor subdirectory).

For both stable and unstable builds, source RPMs are published
to the same locations in artifactory, except under a "src"
directory rather than "noarch" (e.g., a stable build would be
published to: "sat-rpms/hpe/stable/sle-15sp3/docs-sat/X.Y/src/")

Test Description:
* I built the RPM and installed it.
* Ran `rpm -ql` on the installed RPM to list the files.
* Ran `rpm -qi` on the installed RPM to view the RPM info.
* Uninstalled the RPM, verified that the directory was removed.
* I created a test "stable" branch to verify that stable builds
  were being published to the expected stable location.
* I created a test "stable" build of the product stream to verify
  that the RPM can be pulled from the stable location.