CRAYSAT-1610: Update name of management NCN CFS config
Beginning with the Alice release of the HPC CSM Software Recipe, the
management NCN CFS configuration will be named
`management-{{recipe.version}}` where `recipe.version` is the version of
the HPC CSM Software Recipe (currently planned to be 23.03). This single
configuration will be applied both during management NCN image
customization and node personalization. As a result its name is being
changed to `management-{{recipe.version}}` in the default bootprep input
file.

Update the SAT documentation to use this new naming convention. For now,
we are assuming the Alice recipe version is going to be 23.03. If that
changes, we will update it in the docs.