STP-3439: Separate install.md information
Pulled upgrade information, uninstall and downgrade information, and
external system information from install.md into their own MD files.
As part of unified docs, the install and upgrade information is
supposed to be separate, and it makes sense to remove the other distinct
information so it's easier for users to find.

Signed-off-by: Jessica Dehn <jessica.dehn@hpe.com>