STP-3447: Add docs-product-manifest.yaml file
Add a manifest that defines the behavior for unifying docs across all
CSM-based products. Add the file `docs-product-manifest.yaml` to the
RPM and substitute in the version of the docs-sat RPM into the file.
It's not clear what this version is really being used for, but we are
adding it here and making sure it matches the version of the docs,
which is generated from a tag.

Test Description:
Pushed and let Jenkins build RPM. Downloaded RPM, extracted with
`rpm2cpio` and `cpio -idvmu`, and then verified inclusion of
`docs-product-manifest.yaml` and the correct version substitution.

Signed-off-by: Jessica Dehn <jessica.dehn@hpe.com>